requirements.txt
flask
pywin32
psutil
pillow
wmi